+++ core/dbt/tests/fixtures/project.py
@@ -14,6 +14,7 @@ from dbt.events.functions import setup_event_logger
 from dbt.context import providers
 from dbt.events.functions import fire_event
 from dbt.events.test_types import IntegrationTestDebug
+from dbt.tests.util import write_file
 
 
 # These are the fixtures that are used in dbt core functional tests
@@ -111,9 +112,7 @@ def dbt_profile_data(unique_schema, database_host):
 @pytest.fixture
 def profiles_yml(profiles_root, dbt_profile_data):
     os.environ["DBT_PROFILES_DIR"] = str(profiles_root)
-    path = os.path.join(profiles_root, "profiles.yml")
-    with open(path, "w") as fp:
-        fp.write(yaml.safe_dump(dbt_profile_data))
+    write_file(yaml.safe_dump(dbt_profile_data), profiles_root, "profiles.yml")
     yield dbt_profile_data
     del os.environ["DBT_PROFILES_DIR"]
 
@@ -137,8 +136,7 @@ def dbt_project_yml(project_root, project_config_update, logs_dir):
     }
     if project_config_update:
         project_config.update(project_config_update)
-    runtime_config_file = project_root.join("dbt_project.yml")
-    runtime_config_file.write(yaml.safe_dump(project_config))
+    write_file(yaml.safe_dump(project_config), project_root, "dbt_project.yml")
 
 
 # Fixture to provide packages as either yaml or dictionary
@@ -155,7 +153,7 @@ def packages_yml(project_root, packages):
             data = packages
         else:
             data = yaml.safe_dump(packages)
-        project_root.join("packages.yml").write(data)
+        write_file(data, project_root, "packages.yml")
 
 
 # Fixture to provide selectors as either yaml or dictionary
@@ -172,7 +170,7 @@ def selectors_yml(project_root, selectors):
             data = selectors
         else:
             data = yaml.safe_dump(selectors)
-        project_root.join("selectors.yml").write(data)
+        write_file(data, project_root, "selectors.yml")
 
 
 # This creates an adapter that is used for running test setup and teardown,
@@ -205,13 +203,13 @@ def write_project_files(project_root, dir_name, file_dict):
 def write_project_files_recursively(path, file_dict):
     for name, value in file_dict.items():
         if name.endswith(".sql") or name.endswith(".csv") or name.endswith(".md"):
-            path.join(name).write(value)
+            write_file(value, path, name)
         elif name.endswith(".yml") or name.endswith(".yaml"):
             if isinstance(value, str):
                 data = value
             else:
                 data = yaml.safe_dump(value)
-            path.join(name).write(data)
+            write_file(data, path, name)
         else:
             write_project_files_recursively(path.mkdir(name), value)