+++ tests/functional/basic/test_copy_uppercase.py
@@ -1,7 +1,6 @@
 import pytest
-import os
 from dbt.tests.tables import TableComparison
-from dbt.tests.util import run_dbt
+from dbt.tests.util import run_dbt, read_file
 
 from tests.functional.basic.test_simple_copy import (
     advanced_incremental_sql,
@@ -63,11 +62,8 @@ def models():
 @pytest.fixture
 def seeds(test_data_dir):
     # Read seed file and return
-    path = os.path.join(test_data_dir, "seed-initial.csv")
-    with open(path, "rb") as fp:
-        seed_csv = fp.read()
-        return {"seed.csv": seed_csv}
-    return {}
+    seed_csv = read_file(test_data_dir, "seed-initial.csv")
+    return {"seed.csv": seed_csv}
 
 
 def test_simple_copy_uppercase(project):